requirements.txt
# Production requirements - minimal dependencies for inference only
torch==2.8.0
torchaudio==2.8.0
torchlibrosa>=0.1.0
scikit-learn>=1.0.0
soundfile>=0.12.0
numpy>=2.2.6
pandas>=2.3.3
matplotlib>=3.10.6
gradio>=5.49.1
fastapi>=0.118.0
uvicorn[standard]>=0.24.0
mlflow>=2.0.0
librosa>=0.10.0 
scipy>=1.7.0
h5py>=3.7.0
tqdm>=4.60.0
azureml-mlflow>=1.48.0
cloudpickle>=2.0.0
dill>=0.3.5
azure-ai-ml>=1.12.0
azure-identity>=1.16.0
pytest==9.0.2
ruff==0.14.9